Refactor GameController: incremental board display and deferred resume
- GameController split into focused helper methods for parameter
  resolution, board creation, restore, generation, and HUD wiring
- Board and arrow views spawn incrementally during generation/restore
  with a per-frame time budget and loading overlay progress
- SaveManager.LoadAsync for non-blocking save file loading
- Deferred resume: save loaded after loading overlay is visible
- Generation and restore loops share consistent frame-budget pattern

+++ Assets/Scripts/View/GameTimerView.cs
@@ -105,8 +105,7 @@ public sealed class GameTimerView : MonoBehaviour
         return $"{secs}.{millis:D3}";
     }
 
-    private static double GetWallTime() =>
-        (double)DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() / 1000.0;
+    private static double GetWallTime() => DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() / 1000.0;
 
     private void OnDestroy()
     {
